.ipynb_checkpoints/requirements-checkpoint.txt
numpy 
pandas 
matplotlib.pyplot 
yfinance 
scipy.stats 
seaborn
tabulate